+++ lib/airbrake/utils.ex
@@ -33,4 +33,51 @@ defmodule Airbrake.Utils do
       do: {k, @filtered_value},
       else: {k, filter(v, filtered_attributes)}
   end
+
+  @doc """
+  Turns tuples into lists for JSON serialization of Airbrake payloads
+  """
+  def detuple(%module{} = struct) do
+    fields = struct |> Map.from_struct() |> detuple()
+    struct(module, fields)
+  end
+
+  def detuple(map) when is_map(map) do
+    Enum.into(map, %{}, fn {k, v} -> {detuple(k), detuple(v)} end)
+  end
+
+  def detuple(list) when is_list(list) do
+    Enum.map(list, &detuple/1)
+  end
+
+  def detuple(tuple) when is_tuple(tuple) do
+    tuple |> Tuple.to_list() |> detuple()
+  end
+
+  def detuple(other) do
+    other
+  end
+
+  @doc """
+  Recursively breaks down structs for JSON serialization of Airbrake payloads
+  """
+  def destruct(%_module{} = struct) do
+    struct |> Map.from_struct() |> destruct()
+  end
+
+  def destruct(map) when is_map(map) do
+    Enum.into(map, %{}, fn {k, v} -> {destruct(k), destruct(v)} end)
+  end
+
+  def destruct(list) when is_list(list) do
+    Enum.map(list, &destruct/1)
+  end
+
+  def destruct(tuple) when is_tuple(tuple) do
+    tuple |> Tuple.to_list() |> destruct() |> List.to_tuple()
+  end
+
+  def destruct(other) do
+    other
+  end
 end